requirements.txt
# PyTorch and core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# Configuration
pyyaml>=6.0

# Metrics and evaluation
torchmetrics>=0.11.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Optional: For better FID/IS calculations
# lpips>=0.1.4  # Uncomment if you want to use LPIPS for diversity score

